refactor(dispatcher): unify dispatcher and dispatcher/v1 packages
