+++ cmd/hatchet-engine/engine/run.go
@@ -21,7 +21,6 @@ import (
 	"github.com/hatchet-dev/hatchet/internal/services/controllers/retention"
 	"github.com/hatchet-dev/hatchet/internal/services/controllers/task"
 	"github.com/hatchet-dev/hatchet/internal/services/dispatcher"
-	dispatcherv1 "github.com/hatchet-dev/hatchet/internal/services/dispatcher/v1"
 	"github.com/hatchet-dev/hatchet/internal/services/grpc"
 	"github.com/hatchet-dev/hatchet/internal/services/health"
 	"github.com/hatchet-dev/hatchet/internal/services/ingestor"
@@ -361,20 +360,6 @@ func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			return fmt.Errorf("could not start dispatcher: %w", err)
 		}
 
-		dv1, err := dispatcherv1.NewDispatcherService(
-			dispatcherv1.WithRepository(sc.V1),
-			dispatcherv1.WithMessageQueue(sc.MessageQueueV1),
-			dispatcherv1.WithLogger(sc.Logger),
-			dispatcherv1.WithDispatcherId(d.DispatcherId()),
-			dispatcherv1.WithAnalytics(sc.Analytics),
-		)
-
-		if err != nil {
-			return fmt.Errorf("could not create dispatcher (v1): %w", err)
-		}
-
-		d.SetDurableCallbackHandler(dv1.DeliverDurableEventLogEntryCompletion)
-
 		// create the event ingestor
 		ei, err := ingestor.NewIngestor(
 			ingestor.WithMessageQueueV1(sc.MessageQueueV1),
@@ -425,7 +410,7 @@ func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			grpc.WithConfig(sc),
 			grpc.WithIngestor(ei),
 			grpc.WithDispatcher(d),
-			grpc.WithDispatcherV1(dv1),
+			grpc.WithDispatcherV1(d.V1()),
 			grpc.WithAdmin(adminSvc),
 			grpc.WithAdminV1(adminv1Svc),
 			grpc.WithLogger(sc.Logger),
@@ -472,7 +457,6 @@ func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			// hang up long-lived subscriber streams first so that GracefulStop does not
 			// block on them until the pod is killed
 			d.CancelStreamSessions()
-			dv1.CancelStreamSessions()
 
 			g := new(errgroup.Group)
 
@@ -808,20 +792,6 @@ func runV1Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			return fmt.Errorf("could not start dispatcher: %w", err)
 		}
 
-		dv1, err := dispatcherv1.NewDispatcherService(
-			dispatcherv1.WithRepository(sc.V1),
-			dispatcherv1.WithMessageQueue(sc.MessageQueueV1),
-			dispatcherv1.WithLogger(sc.Logger),
-			dispatcherv1.WithDispatcherId(d.DispatcherId()),
-			dispatcherv1.WithAnalytics(sc.Analytics),
-		)
-
-		if err != nil {
-			return fmt.Errorf("could not create dispatcher (v1): %w", err)
-		}
-
-		d.SetDurableCallbackHandler(dv1.DeliverDurableEventLogEntryCompletion)
-
 		// create the event ingestor
 		ei, err := ingestor.NewIngestor(
 			ingestor.WithMessageQueueV1(sc.MessageQueueV1),
@@ -873,7 +843,7 @@ func runV1Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			grpc.WithConfig(sc),
 			grpc.WithIngestor(ei),
 			grpc.WithDispatcher(d),
-			grpc.WithDispatcherV1(dv1),
+			grpc.WithDispatcherV1(d.V1()),
 			grpc.WithAdmin(adminSvc),
 			grpc.WithAdminV1(adminv1Svc),
 			grpc.WithLogger(sc.Logger),
@@ -920,7 +890,6 @@ func runV1Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 			// hang up long-lived subscriber streams first so that GracefulStop does not
 			// block on them until the pod is killed
 			d.CancelStreamSessions()
-			dv1.CancelStreamSessions()
 
 			g := new(errgroup.Group)
