+++ internal/services/dispatcher/server_v1_durable_task_test.go
@@ -1,4 +1,4 @@
-package v1
+package dispatcher
 
 import (
 	"context"
