+++ internal/services/dispatcher/subscribed_worker.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+
 	"github.com/hatchet-dev/hatchet/internal/services/shared/timeout_lock"
 
 	"github.com/hatchet-dev/hatchet/internal/msgqueue"
@@ -11,17 +12,11 @@ import (
 )
 
 type subscribedWorker struct {
-	// stream is the server side of the RPC stream
-	stream contracts.Dispatcher_ListenServer
-
-	// finished is used to signal closure of a client subscribing goroutine
-	finished chan<- bool
-
-	sendLock *timeout_lock.TimeoutLock
-
-	workerId uuid.UUID
-
+	stream    contracts.Dispatcher_ListenServer
+	finished  chan<- bool
+	sendLock  *timeout_lock.TimeoutLock
 	pubBuffer *msgqueue.MQPubBuffer
+	workerId  uuid.UUID
 }
 
 func newSubscribedWorker(
